bakcontent: evolved concept and new implementation
The snapshot and store is now entirely independent.

Snapshot records the state of all files in a git repository. But the
content is not stored, only SHA-512.

Store copy overs the content of all files to supported stores (local
directory only for now) using the SHA-512 of the content as filename.

Store and snapshot can happen in any order, and even concurrently. And
there is no synchronization between the 2 operations.

The only exception is that store will try to use the information
available in the snapshot to avoid having to compute the SHA-512 when
the ctime and mtime of the file are identical to the ones stored in the
snapshot.

Next, we will add systemd files to run regular snapshots on the
bakcontent registered in each user's ~/.bakcontentrc/list. And to run
regular stores.

Pretty good for one day of work.