bakcontent-do-files: lazy_sha512: do not use ctime, only mtime and size
We trust mtime to infer that the content of the file has not changed. By
principle, we ignore files which mtime have been reset.

And if there was a race with a writer (tic; writing; snapshot; writing
(done); tic), so be it.

ctime is still taken into account in the meta file, and if ctime alone
changes, we will record it in an update to the meta file. But we will be
doing so assuming the SHA-512 remains unchanged.

Permissions changes affect ctime which would trigger a lot of work to
recompute hashes without a good reason for it.